Refactor cowsay logic to be used separately
The main part of this refactor is the creation of `fortune-cowsay` which
is a standalone function which can be called interactively, and will
default to copying the output to the kill ring. The secondary part of
this refactor is to use `combine-and-quote-strings` instead of the prior
hacky logic with `mapconcat`. Unfortunately, this is a slight breaking
change in that the custom variables for the arguments now take repeated
strings instead of a single string. So, `'("-f" "tux")` instead of
`"-f tux"`.